changed tack for user login
