+++ package.json
@@ -7,6 +7,7 @@
     "@types/node": "12.0.0",
     "@types/react": "^16.8.17",
     "@types/react-dom": "16.8.4",
+    "babel-polyfill": "^6.26.0",
     "body-parser": "^1.19.0",
     "connect-redis": "^3.4.1",
     "cookie-parser": "^1.4.4",
@@ -21,6 +22,8 @@
     "method-override": "^3.0.0",
     "mongo": "^0.1.0",
     "mongoose": "^5.5.8",
+    "oidc-client": "^1.7.1",
+    "openid-client": "^3.1.1",
     "path": "^0.12.7",
     "querystring": "^0.2.0",
     "react": "^16.8.6",
